pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pj_assistant"
version = "0.1.0"
dependencies = ["playwright", "PyYAML"]
requires-python = ">=3.10"
